_The Great Refactor_: splits tasks into separate files in 'tasks' folder, and makes their structure forced; now called the 'jbfame' module which contains the 'experiment' and 'data' (and optionally utils) scripts; separates related functions and stuff into separate files; renaming galore' includes 2 new tasks based on 'Wei et al' (wiki and base_64).
