+++ experiments/jbfame/jbfame/__init__.py
@@ -1,0 +1,5 @@
+import jbfame.data as data
+import jbfame.experiment as experiment
+import jbfame.utils as utils
+
+__all__ = ["data", "experiment", "utils"]
