+++ experiments/jbfame/jbfame/utils/safety.py
@@ -1,0 +1,74 @@
+from datasets import DatasetDict
+import torch
+
+from jbfame.utils import models 
+from jbfame.utils.params import CustomArguments
+
+
+def _tag_conversation_safety_llama(
+    dataset: DatasetDict, unsafe_column_name: str
+) -> DatasetDict:
+    """
+    Tag questions questions and prompts as 'unsafe' (i.e. unsafe to ask or
+    successful jailbreaks).  The rows should be in the 'chat' format. In case
+    only questions have to be judged, should only contain the 'user' 
+
+    NOTE: Separate function, since purple-llama model needs to be initialized in
+    order to respond.
+    """
+    def is_unsafe(row):
+        if not "chat" in row:
+            row["chat"] = [ {"role": "user", "content": row["prompt"]} ]
+
+        # move to gpu by default. No chance you're gonna run llama 7b on cpu ;)
+        input_ids: torch.Tensor = tokenizer.apply_chat_template(row["chat"], return_tensors="pt").to("cuda") # type: ignore
+
+        output = model.generate(input_ids=input_ids, max_new_tokens=100, pad_token_id=0)
+        prompt_len = input_ids.shape[-1]
+
+        row[unsafe_column_name] = tokenizer.decode(output[0][prompt_len:], skip_special_tokens=True) != 'safe'
+        return row
+
+    model, tokenizer = models.load("meta-llama/LlamaGuard-7b")
+
+    # as we only infer, we do not need the gradients
+    with torch.no_grad():
+        dataset_filtered = dataset.map(is_unsafe)  
+
+    models.clean(model, tokenizer)
+
+    return dataset_filtered
+
+def tag_prompt_safety(
+    datasets: DatasetDict,
+    args: CustomArguments 
+) -> DatasetDict:
+    if not args.model_path.startswith("meta-llama"):
+        # TODO implement a filter for non-llama models
+        raise NotImplementedError("Only llama models are supported for question safety tagging.")
+
+    datasets_tagged = _tag_conversation_safety_llama(datasets, unsafe_column_name="unsafe")
+    
+    return datasets_tagged
+
+def tag_prompt_jailbreak(
+    datasets: DatasetDict,
+    args: CustomArguments 
+) -> DatasetDict:
+    if not args.model_path.startswith("meta-llama"):
+        # TODO implement a filter for non-llama models
+        raise NotImplementedError("Only llama models are supported for question prompt jailbreak tagging.")
+
+    if "unsafe" not in datasets["null"].column_names:
+        datasets_tagged = tag_prompt_safety(datasets, args)
+    else:
+        datasets_tagged = datasets
+
+    # only keep the unsafe questions, and check whether they illicit a jailbreak
+    datasets_tagged = datasets_tagged.filter(lambda row: row["unsafe"])
+    datasets_tagged = _tag_conversation_safety_llama(
+        models.batch_respond(datasets_tagged, args.model_path), 
+        unsafe_column_name="jailbreak"
+    )
+
+    return datasets_tagged
